+++ changelog.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [0.15.0] - 2020-09-08
+### Added
+* Support of JSON format for config files
+* Support of Global Parameters (#29)
+### Changed
+* Function GetObjectsByFolderName (Adf class) uses LIKE operator (#31)
+
 ## [0.14.0] - 2020-07-26
 ### Changed
 * Removed workaround for deployment of Azure Integration Runtimes.